% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IEX_Market_Data.R
\docType{data}
\name{crypto_symbols}
\alias{crypto_symbols}
\title{`IEX` Crypto currency symbols}
\format{
A data frame with 18 rows and 1 variable:
\describe{
  \item{symbol}{symbol, in ticker format}
  ...
}
}
\source{
\url{https://iexcloud.io/docs/api/#cryptocurrency-symbols}
}
\usage{
crypto_symbols
}
\description{
A List of 18 Cyrpto Currency Symbols available via IEX API v1 as of April 26, 2019. List subject to update. Please visit IEX Site for most current list \url{https://iexcloud.io/docs/api/#reference-data}
IEX API version 1 will sunset non-Investors Exchange (IEX) data on June 1, 2019.
}
\keyword{datasets}
