% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IEX_Market_Data.R
\name{iex.account}
\alias{iex.account}
\alias{iex.usage}
\alias{iex.key.usage}
\alias{iex.daily.usage}
\alias{iex.monthly.usage}
\title{Account and usage details}
\usage{
iex.account(iex_sk)

iex.usage(iex_sk)

iex.key.usage(iex_sk)

iex.daily.usage(iex_sk)

iex.monthly.usage(iex_sk)
}
\arguments{
\item{iex_sk}{is a character vector that include IEX Cloud API Secret Token}
}
\value{
Most Active Stocks in IEX during the trading day
}
\description{
For more details, visit:\url{https://iexcloud.io/docs/api/#account}
}
\seealso{
Investors Exchange `IEX` developer guide \url{https://iexcloud.io/docs/api/}

Investors Exchange Group (IEX Group) offers flexible and salable pricing.\url{https://iexcloud.io/pricing/}

View Investors Exchange Group (IEX Group) terms of use and subscription levels.\url{https://iexcloud.io/terms/}

Package `iexcloudR`\url{https://github.com/schardtbc/iexcloudR}
}
\author{
Myriam Ibrahim
}
