% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IEX_Market_Data.R
\name{iex.most.active}
\alias{iex.most.active}
\title{Returns the quotes for the top 10 most active symbols. Data scheduled for intraday update.}
\usage{
iex.most.active(iex_sk)
}
\arguments{
\item{iex_sk}{is a character vector that include IEX Cloud API Secret Token}
}
\value{
Most Active Stocks in IEX during the trading day
}
\description{
For more details, visit:\url{https://iexcloud.io/docs/api/#list}
}
\seealso{
Investors Exchange `IEX` developer guide \url{https://iexcloud.io/docs/api/}

Investors Exchange Group (IEX Group) offers flexible and salable pricing.\url{https://iexcloud.io/pricing/}

View Investors Exchange Group (IEX Group) terms of use and subscription levels.\url{https://iexcloud.io/terms/}

Package `iexcloudR`\url{https://github.com/schardtbc/iexcloudR}
}
\author{
Myriam Ibrahim
}
